feat: :sparkles: User profile fetching and some code restyling
Now it is possible to fetch user informations using the graphql API. Restyling of the leetcode fetch functions. The application is able to check if the user sessions is still active before attempting the login with already existing credentials. Also added credential validations when saving or showing sensitive state informations. Introduced the Table pretty printer and some more colors to the cli.